+++ cmd/buildah/common_unsupported.go
@@ -1,0 +1,5 @@
+// +build !linux,!darwin
+
+package main
+
+func checkUmask() {}
