+++ pkg/parse/parse_unsupported.go
@@ -1,0 +1,7 @@
+// +build !linux,!darwin
+
+package parse
+
+func getDefaultProcessLimits() []string {
+	return []string{}
+}
